[Android] use maxWidth/maxHeight to scale the image (#16)
* Update android code to correctly crop and scale the image

* Fix argument types

* Fix missing files
