+++ .gitignore
@@ -24,3 +24,4 @@ idea
 npm-debug.log
 node_modules/
 bower_components/
+*.iml
